Add tests for CommentReaction, likeit and dislikeit reactions.
